+++ raster/src/main/scala/geotrellis/raster/package.scala
@@ -17,6 +17,7 @@
 package geotrellis
 
 import geotrellis.macros.{NoDataMacros, TypeConversionMacros}
+import geotrellis.vector.{Geometry, Point}
 import spire.math.Integral
 
 
@@ -183,6 +184,7 @@ package object raster extends Implicits {
   implicit class DoubleArrayFiller(val arr: Array[Double]) extends AnyVal {
     def fill(v: Double) = { java.util.Arrays.fill(arr, v) ; arr }
   }
+
 
   /* http://stackoverflow.com/questions/3508077/how-to-define-type-disjunction-union-types */
   sealed class TileOrMultibandTile[T]